Improve the flexibility of `standardize_dtype` and fix `pad` in torch backend (#828)
* improve flexibility in dtype check for torch

* Update

* Fix bugs

* Update

* Fix padding for torch backend

* Update

+++ keras_core/backend/numpy/numpy.py
@@ -216,7 +216,7 @@ def diagonal(x, offset=0, axis1=0, axis2=1):
 
 
 def digitize(x, bins):
-    return np.digitize(x, bins)
+    return np.digitize(x, bins).astype(np.int32)
 
 
 def dot(x, y):
